Upload-4
Add RecyclerView

+++ Expenses/app/src/main/java/com/example/expenses/fragments/Data/Expense.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class Expense(
 
     val explanation: String,
-    val expense: Int,
+    val expense: Double,
     val type: Int,
     val currency: String,
     @PrimaryKey(autoGenerate = true)
